+++ App/common/DatePicker.js
@@ -2,7 +2,6 @@ import React from "react";
 import { StyleSheet } from "react-native";
 import DateTimePicker from "@react-native-community/datetimepicker";
 import colors from "../config/colors";
-import moment from "moment";
 
 const styles = StyleSheet.create({
   date: {
@@ -14,7 +13,7 @@ const styles = StyleSheet.create({
 export default function DatePicker({ dateStyle, onChange, date }) {
   return (
     <DateTimePicker
-      value={moment(date).toDate()}
+      value={date}
       mode="date"
       is24Hour={true}
       display="spinner"
